refactor: move config and chain related stuff to packages
Since they are needed in different contexts, it makes it
easier to move them out of the cmd package and have them
accessible indipendetnly
